pyproject.toml
[project]
name = "nlp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.5.0",
    "gensim>=4.4.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "nltk>=3.9.2",
    "pandas>=2.3.3",
    "pip>=25.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sklearn-crfsuite>=0.5.0",
    "spacy>=3.8.11",
    "transformers>=4.57.3",
]
